+++ SmartIntranet.DTO/DTOs/ReportEmployeeDto/ExcellTemplateModel.cs
@@ -8,7 +8,7 @@ namespace SmartIntranet.DTO.DTOs.ReportEmployeeDto
         public string Position { get; set; }
         public int VacDay { get; set; }
         public int TotalDay { get; set; }
-        public int TotalHour { get; set; }
+        public double TotalHour { get; set; }
         public List<DayItem> DayList { get; set; }
     }
 }
