+++ SmartIntranet.DTO/DTOs/WorkCalendarDto/WorkCalendarListView.cs
@@ -6,7 +6,7 @@ namespace SmartIntranet.DTO.DTOs.WorkCalendarDto
     {
         public string Month { get; set; }
         public int TotalDay { get; set; }
-        public int TotalHour { get; set; }
+        public double TotalHour { get; set; }
         public List<DayType> DayList { get; set; }
     }
 }
