+++ SmartIntranet.DTO/DTOs/WorkGraphicDto/WorkGraphicAddDto.cs
@@ -7,13 +7,13 @@ namespace SmartIntranet.DTO.DTOs.WorkGraphicDto
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Monday { get; set; }
-        public int Tuesday { get; set; }
-        public int Wednesday { get; set; }
-        public int Thursday { get; set; }
-        public int Friday { get; set; }
-        public int Saturday { get; set; }
-        public int Sunday { get; set; }
+        public double Monday { get; set; }
+        public double Tuesday { get; set; }
+        public double Wednesday { get; set; }
+        public double Thursday { get; set; }
+        public double Friday { get; set; }
+        public double Saturday { get; set; }
+        public double Sunday { get; set; }
         public DateTime WorkStart { get; set; }
         public DateTime WorkEnd { get; set; }
         public string Description { get; set; }
