+++ SmartIntranet.Entities/Concrete/IntraHr/WorkGraphic.cs
@@ -8,13 +8,13 @@ namespace SmartIntranet.Entities.Concrete.IntraHr
     {
         public string Name { get; set; }
         public string Key { get; set; }
-        public int Monday { get; set; }
-        public int Tuesday { get; set; }
-        public int Wednesday { get; set; }
-        public int Thursday { get; set; }
-        public int Friday { get; set; }
-        public int Saturday { get; set; }
-        public int Sunday { get; set; }
+        public double Monday { get; set; }
+        public double Tuesday { get; set; }
+        public double Wednesday { get; set; }
+        public double Thursday { get; set; }
+        public double Friday { get; set; }
+        public double Saturday { get; set; }
+        public double Sunday { get; set; }
         public DateTime WorkStart { get; set; }
         public DateTime WorkEnd { get; set; }
         public string Description { get; set; }
